Core/Misc: Update money from 32 bit to 64 bit, increasing maximum money to 1 milion (cataclysm value)

+++ src/server/game/AuctionHouse/AuctionHouseMgr.cpp
@@ -70,7 +70,7 @@ AuctionHouseObject * AuctionHouseMgr::GetAuctionsMap(uint32 factionTemplateId)
         return &mNeutralAuctions;
 }
 
-uint32 AuctionHouseMgr::GetAuctionDeposit(AuctionHouseEntry const* entry, uint32 time, Item *pItem, uint32 count)
+uint64 AuctionHouseMgr::GetAuctionDeposit(AuctionHouseEntry const* entry, uint32 time, Item *pItem, uint32 count)
 {
     uint32 MSV = pItem->GetProto()->SellPrice;
