+++ src/server/game/AuctionHouse/AuctionHouseMgr.h
@@ -60,7 +60,7 @@ struct AuctionEntry
     uint64 owner;
     uint64 startbid;                                        //maybe useless
     uint64 bid;
-    uint32 buyout;
+    uint64 buyout;
     time_t expire_time;
     uint64 bidder;
     uint32 deposit;                                         //deposit can be calculated only when creating auction
@@ -154,7 +154,7 @@ class AuctionHouseMgr
         void SendAuctionOutbiddedMail(AuctionEntry * auction, uint64 newPrice, Player* newBidder, SQLTransaction& trans);
         void SendAuctionCancelledToBidderMail(AuctionEntry* auction, SQLTransaction& trans);
 
-        static uint32 GetAuctionDeposit(AuctionHouseEntry const* entry, uint32 time, Item *pItem, uint32 count);
+        static uint64 GetAuctionDeposit(AuctionHouseEntry const* entry, uint32 time, Item *pItem, uint32 count);
         static AuctionHouseEntry const* GetAuctionHouseEntry(uint32 factionTemplateId);
 
     public:
