+++ src/server/game/Chat/Commands/Level3.cpp
@@ -7121,7 +7121,7 @@ bool ChatHandler::HandleSendMoneyCommand(const char *args)
         return false;
 
     char* money_str = strtok(NULL, "");
-    int32 money = money_str ? atoi(money_str) : 0;
+    int64 money = money_str ? atoll(money_str) : 0;
     if (money <= 0)
         return false;
