+++ src/server/game/Entities/Pet/Pet.cpp
@@ -1668,7 +1668,7 @@ bool Pet::resetTalents(bool no_cost)
         return false;
     }
 
-    uint32 cost = 0;
+    uint64 cost = 0;
 
     if (!no_cost)
     {
@@ -1831,7 +1831,7 @@ void Pet::InitTalentForLevel()
         owner->ToPlayer()->SendTalentsInfoData(true);
 }
 
-uint32 Pet::resetTalentsCost() const
+uint64 Pet::resetTalentsCost() const
 {
     int64 days = int64(sWorld->GetGameTime() - m_resetTalentsTime)/DAY;
